oauth: Add issuer claim
This claim is specifically the underlying/ultimate identity issuer, so
not the federating issuer like the Dex instance the public good instance
uses.

Exposing this is useful for multiple reasons:
* This ultimate issuer is what is needed when verifying the signature
* It's good for users to really see who they are signing as
* Applications that sign may know which identity is required
  so can validate that the identity/issuer are correct _before signing_

It should be noted that this does depend on a bit of a Dex implementation
detail: I don't think there is a standard for exposing the federated
issuer in the token. I consider this similar to how the issuer claim
choice is a bit fuzzy.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>